[hardpy] correct small proplems
