+++ docs/img/dialog_box/base_dialog_box.png
