+++ docs/img/dialog_box/num_input_dialog_box.png
