+++ docs/img/dialog_box/text_input_dialog_box.png
