+++ hardpy/pytest_hardpy/pytest_call.py
@@ -233,7 +233,7 @@ def run_dialog_box(dialog_box_data: DialogBox) -> Any:
         - TEXT_INPUT: str.
         - NUMERIC_INPUT: float.
         - RADIOBUTTON: str.
-        - CHECKBOX: str.
+        - CHECKBOX: list[str].
 
     Raises:
         ValueError: If the 'message' argument is empty.
